+++ string.go
@@ -42,16 +42,13 @@ func (b *StringBuilder[T]) WithDefault(v T) *StringBuilder[T] {
 	return b
 }
 
-// WithConstraintFunc adds a constraint function to the variable.
+// WithConstraint adds a constraint to the variable.
 //
 // fn is called with the environment variable value after it is parsed. If fn
-// returns an error the value is considered invalid.
-//
-// Care should be taken never to include the value in the error message, as it
-// may contain sensitive information.
-func (b *StringBuilder[T]) WithConstraintFunc(
+// returns false the value is considered invalid.
+func (b *StringBuilder[T]) WithConstraint(
 	desc string,
-	fn func(T) variable.ConstraintError,
+	fn func(T) bool,
 ) *StringBuilder[T] {
 	b.spec.UserConstraint(desc, fn)
 	return b